+++ CHANGELOG.md
@@ -34,11 +34,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [\#84](https://github.com/Finschia/finschia-js/pull/84) Bumpup cosmjs to 0.30.1
 - [\#82](https://github.com/Finschia/finschia-js/pull/82) prepare open (change lbm to finschia)
+- [\#90](https://github.com/Finschia/finschia-js/pull/90) change lbmjs-types to finschia-proto
 
 ### Fixed
 
 - [\#73](https://github.com/Finschia/finschia-js/pull/73) fix the local unit test error
 - [\#88](https://github.com/Finschia/finschia-js/pull/88) apply changed event in finschia v1.0.0-rc6 to pass tests
+- [\#91](https://github.com/Finschia/finschia-js/pull/91) ci fix not to pusblish on release
 
 
 ## [v0.7.2]